requirements.txt
openai~=1.47.1
python-dotenv~=1.0.1